feat(sdks): extract boxlite-ffi  from  the C SDK.
