+++ sdks/boxlite-ffi/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "boxlite-ffi"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+description = "Shared FFI layer for BoxLite SDKs"
+
+[lib]
+crate-type = ["rlib"]
+
+[features]
+default = []
+gvproxy-backend = ["boxlite/gvproxy-backend"]
+libslirp-backend = ["boxlite/libslirp-backend"]
+
+[dependencies]
+boxlite = { path = "../../boxlite" }
+
+tokio = { version = "1.37", features = ["rt", "rt-multi-thread"] }
+serde_json = "1.0"
+futures = "0.3"
+
+[dev-dependencies]
+tempfile = "3.10"
+
