+++ sdks/c/Cargo.toml
@@ -12,15 +12,11 @@ crate-type = ["cdylib", "staticlib"]
 
 [features]
 default = []
-gvproxy-backend = ["boxlite/gvproxy-backend"]
-libslirp-backend = ["boxlite/libslirp-backend"]
+gvproxy-backend = ["boxlite-ffi/gvproxy-backend"]
+libslirp-backend = ["boxlite-ffi/libslirp-backend"]
 
 [dependencies]
-boxlite = { path = "../../boxlite" }
-
-tokio = { version = "1.37", features = ["rt", "rt-multi-thread"] }
-serde_json = "1.0"
-futures = "0.3"
+boxlite-ffi = { path = "../boxlite-ffi" }
 
 [build-dependencies]
 cbindgen = "0.29"
