+++ sdks/c/build.rs
@@ -9,17 +9,18 @@ fn main() {
     std::fs::create_dir_all(output_file.parent().unwrap())
         .expect("Failed to create include directory");
 
-    // Generate C header from Rust code
+    // Load cbindgen configuration from cbindgen.toml
+    let config_path = PathBuf::from(&crate_dir).join("cbindgen.toml");
+    let config = cbindgen::Config::from_file(&config_path).expect("Failed to load cbindgen.toml");
+
+    // Generate C header from Rust code (including boxlite-ffi types via parse_deps)
     cbindgen::Builder::new()
         .with_crate(&crate_dir)
-        .with_language(cbindgen::Language::C)
-        .with_pragma_once(true)
-        .with_include_guard("BOXLITE_H")
-        .with_documentation(true)
-        .with_cpp_compat(true)
+        .with_config(config)
         .generate()
         .expect("Unable to generate C bindings")
         .write_to_file(&output_file);
 
     println!("cargo:rerun-if-changed=src/");
+    println!("cargo:rerun-if-changed=cbindgen.toml");
 }
